Add `make_fold` helper to create Folds from function pointers
While Folds already are constructible from function pointers, the naive
apporach of calling the constructor with the two pointers leads to
a stateful fold of size 2*sizeof(void*).

The helper template instead creates a fold-like storing the function
pointers as non-type template arguments, thus yielding a fold which is
an empty class of size 1.