Make numeric operators of `Fold`s free functions
For some reason the concepts branch of clang tries to instantiate the
`lift` function template at the point where it parses the hidden friend
versions of the operators, but it isn't defined there.

To allow building with the concepts branch of clang for now the
operators are free functions again.