Fetch docker image before first build
Right now the docker image is fetched when building with clang. This
has the disadvantage that when looking at the overview provided by
github it seems that clang takes over a minute longer to build than
gcc. In order to detect possible compilation time regressions we
now fetch the image first so that both gcc and clang build times
only include cmake configuration and the actual build.