Add python script to simplify development with docker container
Since the docker commands required to build or test Beautifolds are
quite long there needs to be some kind of helper to make using the
docker container more convenient during interactive development.

This commit adds a python script that allow specifying the action
(build, test, coverage) one would like to perform as well as with
which compiler. Further, to aid in developing new docker based
infrastructure scripts there is also a `run-interactive` action
that drops the user directly into the docker container so that
the necessary commands can be tried interactively before being
put into a script.